Add files via upload
This is my final project for class. I made an Attack/Defend the Castle simulation where the user chooses how they would like to attack the castle whether by foot, boat, or tank. The defenders respond by going to the armory to get arrows and then head to defensive positions. I used four design patterns for this program and they were (in this order) Command, Factory, Decorator, and State. In the main string, I identified where each design pattern is placed in the code.